requirements.txt
google-generativeai
duckduckgo-search
faster-whisper
ctranslate2
edge-tts
faiss-cpu
sentence-transformers
sounddevice
scipy
numpy
pyaudio
pydub
openai-whisper
Pillow
pycaw
pyperclip

# FastAPI Backend
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
pydantic
python-multipart
python-jose
passlib
bcrypt

# Desktop App (PyQt6)
PyQt6
PyQt6-sip

# Database & Storage
# sqlite3 is built-in with Python

# API & Network
requests
urllib3

# Development
python-dotenv